Legal_Knowledge_Base/requirements.txt
streamlit
psycopg2-binary
pymupdf
langchain-openai
pydantic
python-dotenv
pandas